% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{pyramid: making population pyramid of demographics}
\usage{
pyramid(female_pop, male_pop)
}
\arguments{
\item{female_pop}{A data frame containing the disaggregated population estimates for females across all ages groups.
considered.}

\item{male_pop}{A data frame containing the disaggregated population estimates for males across all ages groups.
considered.}
}
\value{
Data frames of the diaggregated population numbers along with uncertainty
for the age and sex groups or other demographic groups.
}
\description{
This function creates population pyramid based on the outputs from the 'cheesecake' or
'cheesepop' functions.
}
\examples{
\dontrun(result <- pyramid(female_pop,male_pop)
}
