% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scissors.R
\name{scissors}
\alias{scissors}
\title{cheesepop: Disaggregating population counts for a single level of demographics
(e.g., age groups only or sex group only) - no geospatial covariates.
Please use 'slices' if you want covariates used.}
\usage{
scissors(df, output_dir)
}
\arguments{
\item{df}{A data frame object containing sample data (often partially observed) on age or sex groups population data
as well as the estimated overall total counts per administrative unit.}

\item{output_dir}{This is the directory with the name of the output folder where the
disaggregated population proportions and population totals are
automatically saved.}
}
\value{
Data frame objects of the output files including the disaggregated population proportions and population totals
along with the corresponding measures of uncertainties (lower and upper bounds of 95-percent credible intervals) for each demographic characteristic.
In addition, a file containing the model performance/model fit evaluation metrics is also produced.
}
\description{
This function disaggregates population estimates by a single demographic (age or sex or religion, etc)
}
\examples{
data(toydata)
result <- scissors(df = toydata, output_dir = tempdir())
}
